+++ src/Symfony/Bundle/SecurityBundle/Tests/DependencyInjection/Fixtures/php/container1.php
@@ -87,11 +87,6 @@ $container->loadFromExtension('security', array(
             'anonymous' => true,
             'http_basic' => true,
         ),
-        'simple_auth' => array(
-            'provider' => 'default',
-            'anonymous' => true,
-            'simple_form' => array('authenticator' => 'simple_authenticator'),
-        ),
     ),
 
     'access_control' => array(